Update Level Scaling Abilities
The 4000 block was temporarily used for 105 Max Level Cap scale for a week on Test, it was changed to 3500 before Live. Some scaling spells still use this, the 1000 range block here is pretty large amount, should be fine to revert back to normal, but the range may be needed to be looked more indepth at.

This also gives extra details on the Max Level Cap that spells allowed to scale to, for spells that Scale by Level (HA and sometimes accidentally used in non-HA content such as Drake Fireballs in Skyfire pre-Fix)